refactor: split client and Metro URL ownership (#2146)
