+++ src/client/base-url.ts
@@ -1,0 +1,7 @@
+export function normalizeBaseUrl(input: string): string {
+  let end = input.length;
+  while (end > 0 && input.charCodeAt(end - 1) === 47) {
+    end -= 1;
+  }
+  return end === input.length ? input : input.slice(0, end);
+}
