+++ src/client/client-companion-tunnel-worker.ts
@@ -25,7 +25,7 @@ import type {
   MetroTunnelRequestMessage as MetroCompanionRequest,
   MetroTunnelResponseMessage,
 } from '../metro/metro.ts';
-import { normalizeBaseUrl } from '../utils/url.ts';
+import { normalizeBaseUrl } from './base-url.ts';
 
 const COMPANION_REGISTER_TIMEOUT_MS = 5_000;
 const COMPANION_REGISTER_MAX_RETRY_DELAY_MS = 60_000;
