+++ src/metro/bundle-url.ts
@@ -1,10 +1,4 @@
-export function normalizeBaseUrl(input: string): string {
-  let end = input.length;
-  while (end > 0 && input.charCodeAt(end - 1) === 47) {
-    end -= 1;
-  }
-  return end === input.length ? input : input.slice(0, end);
-}
+import { normalizeBaseUrl } from '../client/base-url.ts';
 
 export function buildBundleUrl(
   baseUrl: string,
